refactor: 피드백 반영. reply 패키지 comment로 합침, 인가 이후 댓글 찾음 (#367)
