+++ src/main/java/com/exchangediary/comment/domain/CommentRepository.java
@@ -1,5 +1,6 @@
-package com.exchangediary.comment.domain.entity;
+package com.exchangediary.comment.domain;
 
+import com.exchangediary.comment.domain.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
