+++ src/main/java/com/exchangediary/comment/domain/ReplyRepository.java
@@ -1,5 +1,6 @@
-package com.exchangediary.reply.domain.entity;
+package com.exchangediary.comment.domain;
 
+import com.exchangediary.comment.domain.entity.Reply;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReplyRepository extends JpaRepository<Reply, Long> {
